Remove invalid save/load state tests.
The tests were failing because they are designed to keep state between
actions.

Hypothesis generates a simple program like::

    state = GCounterComparison()
    v1 = state.replica(name=0)
    state.run_incr(replica=v1)
    state.load_state()
    state.run_synchronize(receiver=v1)

Notice that variable `v1` is the replicated object and that it is not affected
by the `load_state` command.  Which means run_synchronize is test the state
before the load_state and compares with the model after the load.