Split concurrent/non-concurrent tests for LWWRegister.
There was a problem with the test-model, it didn't actually allowed any
concurrent update.

`run_set` will update the test-model without any regards to the possible state
of the whole replica-set.

I'm splitting the concurrent from the non-concurrent machinery.  Because using
a strictly monotonic time function allows for this simple test-model:

    Time     Replica 1      Replica 2  ...     Model
    1        set to x                          x
    2                       set to y           y

There's no problem to update the Model to y, because Time is total, even
though the vector clocks of the replicas are concurrent ([1, 0] and [0, 1])
their timestamp makes Replica 2 a winner.

The problem happens with:

    Time     Replica 1      Replica 2  ...     Model
    1        set to x       set to y           ?

What is the right value for Model?