Simplify LWWRegister concurrence tests.
Two consecutive runs of `run_possibly_concurrent_set` may lead to concurrent
updates in different replicas.  We only update the replica if the test-model
is updated.

Assume the following scenario:

   Time   R0           R1           R2           Model
   0      set to 1                               1
   0                   set to 2                  2

At this moment synchronize R2 merging R0 first: 'R2 << R1' is True because
R2's vclock is empty.  But after the merge, R2's is concurrent with R1 and 'R1
<< R2' so the value of R1 is discarded; R2 keeps the value 1 but R2 is now a
descedant of R1 and R0.

If we now merge R1 with R2 it will adjust it's value to 1.

So I think, the way to test LWWRegister is to perform a full-synchronization
and test that all replicas reached the same value.